Transfer getCurrentUser method from AvailabilityService to UserService
