+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Interfaces/IUserService.cs
@@ -1,0 +1,9 @@
+﻿using EasyMeets.Core.Common.DTO;
+
+namespace EasyMeets.Core.BLL.Interfaces
+{
+    public interface IUserService
+    {
+        public Task<UserDto> GetCurrentUserForAvailabilityPageAsync(int id);
+    }
+}
