+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Services/AvailabilityService.cs
@@ -43,14 +43,7 @@ namespace EasyMeets.Core.BLL.Services
                  .ToListAsync();
             var userSlotsDto = _mapper.Map<ICollection<AvailabilitySlotDto>>(userSlots);
             return userSlotsDto;
-        }
-
-        public async Task<UserDto> GetCurrentUserForAvailabilityPageAsync(int id)
-        {
-            var currentUser = await _context.Users.FirstOrDefaultAsync(x => x.Id == id);
-            var currentUserDto = _mapper.Map<UserDto>(currentUser);
-            return currentUserDto;
-        }
+        } 
 
         public async Task CreateAvailabilitySlot(NewAvailabilitySlotDto slotDto)
         {
