+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Services/UserService.cs
@@ -1,0 +1,20 @@
+﻿using AutoMapper;
+using EasyMeets.Core.BLL.Interfaces;
+using EasyMeets.Core.Common.DTO;
+using EasyMeets.Core.DAL.Context;
+using Microsoft.EntityFrameworkCore;
+
+namespace EasyMeets.Core.BLL.Services
+{
+    public class UserService : BaseService, IUserService
+    {
+        public UserService(EasyMeetsCoreContext context, IMapper mapper) : base(context, mapper) {}
+
+        public async Task<UserDto> GetCurrentUserForAvailabilityPageAsync(int id)
+        {
+            var currentUser = await _context.Users.FirstOrDefaultAsync(x => x.Id == id);
+            var currentUserDto = _mapper.Map<UserDto>(currentUser);
+            return currentUserDto;
+        }
+    }
+}
